pyproject.toml
[tool.poetry]
name = "test-1"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.11,<4.0"

# these were available in legacy repls. might be helpful to add
# glfw = "2.2.0"
# toml = "0.10.2"
# importlib-metadata = "3.10.1"
# parso = "0.5.2"
# pluggy = "0.13.1"
# typing-extensions = "3.7.4.3"
# ujson = "1.35"
# zipp = "3.4.1"
# jedi = "0.15.2"
# pyflakes = "2.1.1"
# rope = "0.18.0"
# yapf = "0.31.0"
# matplotlib = "3.4.3"
# nltk = "3.6.2"
# numpy = "^1.21.5"
# scipy = "1.7.3"
ffmpeg-python = "^0.2.0"
flask = "^3.1.2"
flask-discord = "^0.1.69"
discord-py = "^2.6.4"
praw = "^7.8.1"
youtube-dl = "^2021.12.17"

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
